. install the timer that will fire the auto-reimporter. . fire the timer using a Swing timer, not a java.util.Timer. This makes it easier to handle a race between the timer thread running the reimporter and the AWT thread trying to change a project; the race is not yet fully handled, but starting the reimport task in the AWT thread makes it a lot easier to do.
The actual reimport will happen on a different thread, since it does I/O.

+++ plugins/ProjectViewer/trunk/projectviewer/config/AutoReimportPane.java
@@ -121,11 +121,11 @@ class AutoReimportPane extends OptionPaneBase
 
 	/** Saves the configuration to the project. */
 	protected void _save() {
-		long pval = 0;
+		int pval = 0;
 
 		if (enable.isSelected()) {
 			try {
-				pval = Long.parseLong(period.getText());
+				pval = Integer.parseInt(period.getText());
 			} catch (NumberFormatException nfe) {
 				Log.log(Log.WARNING, this,
 						"Invalid auto import period: " + period.getText());
